feat(configtx-cli): Add initial implementation of Fabric ConfigTX CLI
This is the first commit of Fabric ConfigTX CLI, which is a tiny
channel / organization management tool for Hyperledger Fabric v2.x.
This tool outputs a config transaction that controls channels and/or
organizations by using input parameters.
This CLI will be internally used by other components of OpsSC.

See `configtx-cli/README.md` for more information.

Signed-off-by: Tatsuya Sato <Tatsuya.Sato@hal.hitachi.com>