fix(common): Add processing to wait for peers to complete joining channel
The current `peer channel join`, to join a peer to a channel, seems
to return success before all processes for joining have done.

Therefore, the peer being joined may not be discovered immediately
after executing this command.

As the workaround, this patch adds processing to wait for peers to
complete joining channel by executing `peer channel list` in
joinAllPeers().

Signed-off-by: Tatsuya Sato <Tatsuya.Sato@hal.hitachi.com>