feat(common): [Preparation] Copy some protobuf handlers from FOC for private data support
As a preparation for supporting chaincode deployment with private data, this
patch copies the relevant protobuf handlers from Fabric Operations Console (FOC).

The source code are copied from:
- https://github.com/hyperledger-labs/fabric-operations-console
- commit: 0bd6cfd35722a649b1b8df778aeadf0daedce556

Note that this patch alone is not sufficient to work, as they are simply copied
to clarify the differences from the original.

Signed-off-by: Tatsuya Sato <tatsuya.sato.so@hitachi.com>