fix(agent): Fix a bug when handling a proposal not using the next sequence
Normally, if a proposal uses an incorrect sequence number, there is no problem
because committing the chaincode definition will fail.

However, in the case of a proposal with the current sequence number and a new
chaincode source code, the approval process will update the chaincode package
in the sequence unexpectedly. Furthermore, in the case of using an earlier
sequence number, resources are wasted due to installing the chaincode successfully.

To avoid the cases, this patch adds a process to compare the committed sequence
number with the proposed one at the top of prepareToDeploy().

Signed-off-by: Tatsuya Sato <tatsuya.sato.so@hitachi.com>