chore(chaincode): Update dependencies
This patch updates dependencies on chaincodes.
The main reason for this is to update the version of
fabric-contract-api-go to the latest v1.2.0.
This version simplifies support for running chaincode
as a service. It automatically starts chaincode in sever
mode if CHAINCODE_SERVER_ADDRESS and CORE_CHAINCODE_ID_NAME
environment variables are set.

Signed-off-by: Tatsuya Sato <tatsuya.sato.so@hitachi.com>