fix(common): Improve error handling for invoking/querying CC
It appears that Fabric node SDK's Gateway may not properly update the
service discovery in some corner cases. For example, if a non-existent
MSP ID is initially configureed in the endorse policy and then the
correct policy is reconfigured, it is not updated.

This patch will force rediscovery in the above cases by closing the
Gateway once anyway when chaincode invoking/querying errors occur in
fabric-client.

Signed-off-by: Tatsuya Sato <tatsuya.sato.so@hitachi.com>